+++ server/graphql/core/src/loader/loader_registry.rs
@@ -341,5 +341,13 @@ pub async fn get_loaders(
         },
         async_std::task::spawn,
     ));
+
+    loaders.insert(DataLoader::new(
+        VaccineCourseByProgramIdLoader {
+            connection_manager: connection_manager.clone(),
+        },
+        async_std::task::spawn,
+    ));
+
     loaders
 }
