+++ server/graphql/core/src/loader/mod.rs
@@ -34,6 +34,7 @@ mod sync_file_reference;
 mod temperature_breach;
 mod temperature_log;
 mod user;
+mod vaccine_course_by_program;
 
 use std::{collections::HashSet, hash::Hasher};
 
@@ -73,6 +74,7 @@ pub use sync_file_reference::*;
 pub use temperature_breach::*;
 pub use temperature_log::*;
 pub use user::*;
+pub use vaccine_course_by_program::*;
 
 #[derive(Debug, Clone)]
 /// Sometimes loaders need to take an extra parameter, like store_id or requisition_id
